feat(editor): implement new editor component with menu and bubble functionalities
