--- src/components/wysiwyg-editor/wysiwyg-editor.stories.tsx
@@ -1,17 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-
-import { WYSIWYGEditor } from './wysiwyg-editor';
-
-const meta: Meta = {
-  component: WYSIWYGEditor,
-  title: 'Components/WYSIWYG Editor',
-};
-
-export default meta;
-type Story = StoryObj<typeof WYSIWYGEditor>;
-
-export const Demo: Story = {
-  render: (_) => {
-    return <WYSIWYGEditor />;
-  },
-};
