+++ src/index.ts
@@ -37,7 +37,7 @@ export * from '@components/breadcrumb/breadcrumb';
 export * from '@components/color-picker-input/color-picker-input';
 export * from '@components/drag-n-drop-file-input/drag-n-drop-file-input';
 export * from '@components/checkbox-tree-view/checkbox-tree-view';
-export * from '@components/wysiwyg-editor/wysiwyg-editor';
+export * from '@components/editor/editor';
 
 // Other stuff
 export * from '@src/lib/icons';
